+++ Adafruit AR/art.scnassets/Settings.json
@@ -1,0 +1,5 @@
+{
+  "interleaveGeometrySources" : true,
+  "forceYUp" : true,
+  "preferCompressedTextures" : true
+}
