+++ Assets.xcassets/Photos.arresourcegroup/Contents.json
@@ -8,6 +8,9 @@
       "filename" : "CPX-QRCode_v2.arreferenceimage"
     },
     {
+      "filename" : "cpx.arreferenceimage"
+    },
+    {
       "filename" : "CrickitQRCode_v2.arreferenceimage"
     },
     {
@@ -18,6 +21,9 @@
     },
     {
       "filename" : "NeoTrellis_v2.arreferenceimage"
+    },
+    {
+      "filename" : "portal2.arreferenceimage"
     },
     {
       "filename" : "pyportal.arreferenceimage"
